+++ src/rust/ficus_backend/src/pipelines/pipelines.rs
@@ -627,7 +627,7 @@ impl PipelineParts {
     fn filter_traces_by_count() -> (String, PipelinePartFactory) {
         Self::create_pipeline_part(Self::FILTER_TRACES_BY_EVENTS_COUNT, &|context, keys, config| {
             let log = Self::get_context_value_mut(context, keys.event_log())?;
-            let min_events_count = *Self::get_context_value(config, keys.events_count())?;
+            let min_events_count = *Self::get_context_value(config, keys.events_count())? as usize;
             log.filter_traces(&|trace, _| trace.get_events().len() < min_events_count);
 
             Ok(())
